Added Light Atack Follow Up
Added Sprites, Flipbook, and Attack Instance for a second light attack after the first.

Updated AttackInstance to use functions ShouldHitActor and HitActor since they will be common to all children.

Need to update State Machine to use it in a new state now.
Need a new naming scheme for sprites.